+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.10.0] - 2018-08-01
+
+### Added
+
++ Added media secret
+
+### Fixed
+
++ Fixed incorrect term in readme
+
 ## [2.9.1] - 2018-07-31
 
 ### Fixed
@@ -287,7 +297,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.9.1...develop
+[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.10.0...develop
+[2.10.0]: https://github.com/mserajnik/hydrusrv/compare/2.9.1...2.10.0
 [2.9.1]: https://github.com/mserajnik/hydrusrv/compare/2.9.0...2.9.1
 [2.9.0]: https://github.com/mserajnik/hydrusrv/compare/2.8.1...2.9.0
 [2.8.1]: https://github.com/mserajnik/hydrusrv/compare/2.8.0...2.8.1